added table and renamed modules
